float output
There are now two http response protocols. One to serve the initial html and then another to request a "float", a float wrapped in a string.

The Deci2 server in the game allows for packets to be sent with inspect calls that enumerate the symbol in all types. By grabbing the float from that output and sending that in place of the hardcoded "float" I'm using now, the float can be sent from the game itself. For instance, (-> *target* root trans x )